Moved cloud config base file to root directory.
Modified vagrant config to save generated cloud config back to old location, which is added to .gitignore.
This prevents every vagrant up from overwriting the committed cloud config version.

Also created "fleet" task, which will execute a fleetctl command against all of the units listed within that task.
This task is used in the Vagrantfile to ensure that all expected units are started on the first vagrant up.