Modified jenkins-master unit file to work around folder permission errors on the jenkins-master-data volume.
Simply modifying the owner of the volume folder doesn't seem to work when no other containers are running against it during the unit creation.
Therefore, run a temporary container with that volume mounted, during the pre-start sequence for the unit. The temporary container is removed after the owner of the volume folder is changed by a separate docker run command.